using as operator to cast string to int

+++ BardDomainModel/grails-app/migrations/reset-sequences.groovy
@@ -12,7 +12,7 @@ databaseChangeLog = {
                     String tableName = sequenceName - '_ID_SEQ'
                     String maxIdQuery = "select max(${tableName}_ID) as max_id from ${tableName}"
                     try {
-                        int maxId = sql.rows(maxIdQuery)[0].max_id ?: 0
+                        int maxId = (sql.rows(maxIdQuery)[0].max_id ?: 0) as int
                         int startWithVal = maxId + 1
 
                         String dropSeqSql = "drop sequence ${sequenceName}"
